applying latest changes of generator to loadbalancer, database and object storage

+++ objectstorage/get_object_request_response.go
@@ -66,7 +66,7 @@ type GetObjectResponse struct {
 	OpcMeta_ *string `presentIn:"header" name:"opc-meta-*"`
 
 	// The object size in bytes.
-	ContentLength *int64 `presentIn:"header" name:"content-length"`
+	ContentLength *int `presentIn:"header" name:"content-length"`
 
 	// Content-Range header for range requests, per [RFC 7233](https://tools.ietf.org/rfc/rfc7233), section 4.2.
 	ContentRange *string `presentIn:"header" name:"content-range"`
