+++ objectstorage/put_object_request_response.go
@@ -23,7 +23,7 @@ type PutObjectRequest struct {
 	ObjectName *string `mandatory:"true" contributesTo:"path" name:"objectName"`
 
 	// The content length of the body.
-	ContentLength *int64 `mandatory:"true" contributesTo:"header" name:"Content-Length"`
+	ContentLength *int `mandatory:"true" contributesTo:"header" name:"Content-Length"`
 
 	// The object to upload to the object store.
 	PutObjectBody *string `mandatory:"true" contributesTo:"body"`
